Split up window event handling
