--- src/platform_impl/windows/event_loop/handler/destroy.rs
@@ -1,3 +1,0 @@
-//! `WM_DESTROY`, `WM_CLOSE` and `WM_NCDESTROY`.
-
-
